refactor: alter request and resource paths
