+++ app/Http/Requests/Task/TaskRequest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Requests;
+namespace App\Http\Requests\Task;
 
 use App\Enums\StatusEnum;
 use Illuminate\Foundation\Http\FormRequest;
